Add lookup, refactor jazzy path
